requirements.txt
# Core Python Packages
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
bitarray==3.3.1
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-default-group==1.2.4
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kombu==5.5.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
namex==0.0.8
narwhals==1.34.1
networkx==3.4.2
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
pyarrow==19.0.1
pybloom-live==4.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv 
fastapi
python-multipart
python-multipart==0.0.20
python-slugify==8.0.4
python-whois==0.9.5
pytz==2025.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rich==14.0.0
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
sqlite-fts4==1.0.3
sqlite-utils==3.38
starlette==0.46.2
streamlit==1.44.1
sympy==1.13.3
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.0.1
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.19.0

# ✅ AI/ML for HuggingFace / URL analysis
sentence-transformers
transformers
torch
xgboost
tf-keras

# ✅ Sandboxing for Phishing URL Testing
playwright
playwright-stealth

beautifulsoup4==4.12.3
lxml==5.2.1  # BeautifulSoup


